# server/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_costrict_costrict-web_server_internal_models.CapabilityArtifact:
    properties:
      artifactVersion:
        type: string
      checksumSha256:
        type: string
      createdAt:
        type: string
      downloadCount:
        type: integer
      fileSize:
        type: integer
      filename:
        type: string
      id:
        type: string
      isLatest:
        type: boolean
      itemId:
        type: string
      mimeType:
        type: string
      sourceType:
        type: string
      storageBackend:
        type: string
      storageKey:
        type: string
      uploadedBy:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityAsset:
    properties:
      contentSha:
        type: string
      createdAt:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      itemId:
        type: string
      mimeType:
        type: string
      relPath:
        type: string
      storageBackend:
        type: string
      storageKey:
        type: string
      textContent:
        type: string
      updatedAt:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityItem:
    properties:
      artifacts:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityArtifact'
        type: array
      assets:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityAsset'
        type: array
      category:
        type: string
      content:
        type: string
      contentMd5:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      currentRevision:
        type: integer
      description:
        type: string
      embeddingUpdatedAt:
        type: string
      evaluation:
        type: object
      experienceScore:
        type: number
      favoriteCount:
        type: integer
      health:
        type: object
      id:
        type: string
      installCount:
        type: integer
      itemType:
        type: string
      lastScanId:
        type: string
      metadata:
        type: object
      name:
        type: string
      previewCount:
        type: integer
      registry:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
      registryId:
        type: string
      repoId:
        type: string
      securityStatus:
        type: string
      slug:
        type: string
      source:
        description: 导入来源，如 anthropic/claude-code, superpower, github
        type: string
      sourcePath:
        type: string
      sourceSha:
        type: string
      sourceType:
        description: direct | archive
        type: string
      status:
        type: string
      tags:
        description: Tags loaded separately (not persisted by GORM)
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.ItemTagDict'
        type: array
      updatedAt:
        type: string
      updatedBy:
        type: string
      version:
        type: string
      versions:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityVersion'
        type: array
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry:
    properties:
      createdAt:
        type: string
      description:
        type: string
      externalBranch:
        type: string
      externalUrl:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityItem'
        type: array
      lastSyncLog:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncLog'
      lastSyncLogId:
        type: string
      lastSyncSha:
        type: string
      lastSyncedAt:
        type: string
      name:
        type: string
      ownerId:
        type: string
      repoId:
        type: string
      sourceType:
        type: string
      syncConfig:
        type: object
      syncEnabled:
        type: boolean
      syncInterval:
        type: integer
      syncStatus:
        description: idle | syncing | error | paused
        type: string
      updatedAt:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.CapabilityVersion:
    properties:
      assets:
        items:
          $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityVersionAsset'
        type: array
      category:
        type: string
      commitMsg:
        type: string
      content:
        type: string
      contentMd5:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      description:
        type: string
      id:
        type: string
      itemId:
        type: string
      metadata:
        type: object
      name:
        type: string
      revision:
        type: integer
      sourcePath:
        type: string
      version:
        type: string
    type: object
...
  github_com_costrict_costrict-web_server_internal_models.ItemTagDict:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      id:
        type: string
      slug:
        type: string
      tagClass:
        description: system | custom
        type: string
    type: object
...
  github_com_costrict_costrict-web_server_internal_models.SyncJob:
    properties:
      createdAt:
        type: string
      finishedAt:
        type: string
      id:
        type: string
      lastError:
        type: string
      maxAttempts:
        type: integer
      payload:
        type: object
      priority:
        type: integer
      registry:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
      registryId:
        type: string
      retryCount:
        type: integer
      scheduledAt:
        type: string
      startedAt:
        type: string
      status:
        description: pending | running | success | failed | cancelled
        type: string
      syncLogId:
        type: string
      triggerType:
        description: scheduled | manual | webhook
        type: string
      triggerUser:
        type: string
    type: object
  github_com_costrict_costrict-web_server_internal_models.SyncLog:
    properties:
      addedItems:
        type: integer
      commitSha:
        type: string
      createdAt:
        type: string
      deletedItems:
        type: integer
      durationMs:
        type: integer
      errorMessage:
        type: string
      failedItems:
        type: integer
      finishedAt:
        type: string
      id:
        type: string
      previousSha:
        type: string
      registry:
        $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.CapabilityRegistry'
      registryId:
        type: string
      skippedItems:
        type: integer
      startedAt:
        type: string
      status:
        description: running | success | failed | cancelled
        type: string
      totalItems:
        type: integer
      triggerType:
        description: scheduled | manual | webhook
        type: string
      triggerUser:
        type: string
      updatedItems:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    name: CoStrict Team
    url: https://costrict.ai
  description: AI Agent Platform API - Skill marketplace, repository and registry
    management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Costrict Web API
  version: "1.0"
paths:
...
  /registries/{id}/sync:
    post:
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      - description: Dry run mode
        in: query
        name: dryRun
        type: boolean
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            properties:
              jobId:
                type: string
              status:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Trigger registry sync
      tags:
      - sync
  /registries/{id}/sync-jobs:
    get:
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              jobs:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncJob'
                type: array
              total:
                type: integer
            type: object
      summary: List registry sync jobs
      tags:
      - sync
  /registries/{id}/sync-logs:
    get:
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              logs:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncLog'
                type: array
              total:
                type: integer
            type: object
      summary: List registry sync logs
      tags:
      - sync
  /registries/{id}/sync-status:
    get:
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Get registry sync status
      tags:
      - sync
  /registries/{id}/sync/cancel:
    post:
      parameters:
      - description: Registry ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      summary: Cancel registry sync
      tags:
      - sync
...
  /repositories/{id}/sync:
    post:
      description: Manually trigger a sync job for the repository's registry
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry ID
        in: query
        name: registryId
        type: string
      - description: Dry run mode
        in: query
        name: dryRun
        type: boolean
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            properties:
              jobId:
                type: string
              status:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Trigger repo sync
      tags:
      - sync
  /repositories/{id}/sync-jobs:
    get:
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry ID (filter by registry)
        in: query
        name: registryId
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              jobs:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncJob'
                type: array
              total:
                type: integer
            type: object
      summary: List repo sync jobs
      tags:
      - sync
  /repositories/{id}/sync-logs:
    get:
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry ID (filter by registry)
        in: query
        name: registryId
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              logs:
                items:
                  $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncLog'
                type: array
              total:
                type: integer
            type: object
      summary: List repo sync logs
      tags:
      - sync
  /repositories/{id}/sync-status:
    get:
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry ID (get specific registry status)
        in: query
        name: registryId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Get repo sync status
      tags:
      - sync
  /repositories/{id}/sync/cancel:
    post:
      parameters:
      - description: Repository ID
        in: path
        name: id
        required: true
        type: string
      - description: Registry ID (cancel specific registry)
        in: query
        name: registryId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
            type: object
      summary: Cancel repo sync
      tags:
      - sync
...
  /sync-jobs/{id}:
    get:
      parameters:
      - description: SyncJob ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncJob'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get sync job detail
      tags:
      - sync
  /sync-logs/{id}:
    get:
      parameters:
      - description: SyncLog ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_costrict_costrict-web_server_internal_models.SyncLog'
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get sync log detail
      tags:
      - sync
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"